drivers/led_driver: Change PWM brightness type to uint8_t

+++ src/drivers/led_driver.h
@@ -114,7 +114,7 @@ void led_driver_irq_handler(void);
   * @param      procent_val: PWM value in [%].
   * @retval     None.
   *****************************************************************************/
-void led_driver_pwm_set_brightness(uint16_t procent_val);
+void led_driver_pwm_set_brightness(uint8_t procent_val);
 
 /*******************************************************************************
   * @function   led_driver_pwm_get_brightness
@@ -122,7 +122,7 @@ void led_driver_pwm_set_brightness(uint16_t procent_val);
   * @param      None.
   * @retval     procent_val: PWM value in [%].
   *****************************************************************************/
-uint16_t led_driver_pwm_get_brightness(void);
+uint8_t led_driver_pwm_get_brightness(void);
 
 /*******************************************************************************
   * @function   led_driver_step_brightness
